requirements.txt
Flask==1.1.2
Flask-Login==0.5.0
flask-talisman==0.7.0
google-cloud-datastore==1.15.0
google-cloud-storage==1.31.2
google-cloud-ndb==1.6.0
gunicorn==20.0.4
twilio==6.45.4
sendgrid==6.4.7
phonenumbers==8.12.10
email_validator==1.1.1
python-dateutil==2.8.1